requirements.txt
asgiref==3.9.1
certifi==2025.7.9
channels==4.3.1
charset-normalizer==3.4.2
Django==5.2.4
django-appconf==1.1.0
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-imagekit==5.0.0
django-jazzmin==3.0.1
django-model-utils==5.0.0
django-rest-passwordreset==1.5.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
pilkit==3.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
setuptools==80.9.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.9.0
